refactor: rename service directories to services, remove redundant service, and ignore daily-log files
