+++ src/lark-meeting/services/lark-event-ws.service.ts
@@ -3,7 +3,7 @@
  * @Date: 2025-11-23 17:56:51
  * @LastEditors: 杨仕明 shiming.y@qq.com
  * @LastEditTime: 2025-11-23 18:23:14
- * @FilePath: /lulab_backend/src/lark-meeting/service/lark-event-ws.service.ts
+ * @FilePath: /nove_api/src/lark-meeting/services/lark-event-ws.service.ts
  * @Description:
  * *
  * Copyright (c) 2025 by LuLab-Team, All Rights Reserved.
