+++ src/task/tasks.controller.ts
@@ -22,7 +22,7 @@ import {
 } from '@nestjs/common';
 import { ApiTags, ApiExtraModels } from '@nestjs/swagger';
 import { NoPermissionRequired } from '@/permission/decorators/permissions.decorator';
-import { TasksService } from './service/tasks.service';
+import { TasksService } from './services/tasks.service';
 import { CreateOnceDto } from './dto/create-once.dto';
 import { CreateCronDto } from './dto/create-cron.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
