+++ src/wechat-shop/wechat-shop.module.ts
@@ -9,10 +9,10 @@ import { BullMQAdapter } from '@bull-board/api/bullMQAdapter';
 import { WechatShopOrderController } from './controllers/wechat-shop-order.controller';
 import { WechatShopEventController } from './controllers/wechat-shop-event.controller';
 import { WechatShopRepository } from './repositories';
-import { WechatShopOrderService } from './service/wechat-shop-order.service';
-import { WechatShopEventService } from './service/wechat-shop-event.service';
-import { WechatShopClientService } from './service/wechat-shop-client.service';
-import { WechatShopTokenService } from './service/wechat-shop-token.service';
+import { WechatShopOrderService } from './services/wechat-shop-order.service';
+import { WechatShopEventService } from './services/wechat-shop-event.service';
+import { WechatShopClientService } from './services/wechat-shop-client.service';
+import { WechatShopTokenService } from './services/wechat-shop-token.service';
 import { WechatShopProcessor } from './processor/wechat-shop.processor';
 import { UserModule } from '@/user/user.module';
